+++ src/main/java/net/pms/util/jna/TerminatedArray.java
@@ -114,7 +114,7 @@ public abstract class TerminatedArray<E> extends PointerType implements List<E> 
 		if (buffer == null) {
 			return null;
 		}
-		long size = getElementSize() * (buffer.size() + 1);
+		long size = (long) getElementSize() * (buffer.size() + 1);
 		// Reuse the reserved memory if possible
 		if (!(getPointer() instanceof Memory) || ((Memory) getPointer()).size() < size) {
 			super.setPointer(new Memory(size));
